qa/tasks: add template.py for generic templating needs
This change extracts a bunch of functionality that we added to
cephadm.py and has been "incubating" there for too long as it's not
particularly specific to cephadm.

Take the code from cephadm.py and fully generalize it and make it a
fairly simple module that can be reused elsewhere as needed.

Signed-off-by: John Mulligan <jmulligan@redhat.com>