doc/cephadm/services: Improve formatting and small fixes in index.rst
Indent block formatting to the left margin of the current paragraph.
And indent with a consistent amount of spaces.
Fix some text paragraphs indented too much.

Remove spaces at the end of lines.
Remove a stray space preceding a comma and in the middle of a hostname.
Add some full stops and colons where suitable.

Use title case in section titles consistently.
Fix some capitalization errors in text: YAML, CPU, some words starting
with capital case in the middle of a sentence.
Always use "service specification" (capitalized normally) instead of
also sometimes "Service Specification" or "Services Specification".
Cephadm the concept was sometimes capitalized even in the middle of a
sentence, sometimes not capitalized even in the beginning of a sentence.
Only capitalize it at the beginning of a sentence.
Capitalize "MON" when it refers to daemon/service and not to a
configuration parameter, for consistency with other such services.

Use double backticks instead of single backtick or otherwise for data
like warning names, CLI commands, file names.

Don't direct to add a space after "regex:" for a pattern, which
contradicted the next sample.

Try to improve a sentence to also mention the host list after the count
in an example command.

Use privileged user prompt consistently for privileged example CLI
commands.

Signed-off-by: Ville Ojamo <14869000+bluikko@users.noreply.github.com>